[refactor] folder reorganization
