+++ src/api/nftRoutes.ts
@@ -5,7 +5,6 @@ const nftRoutes = async (fastify: FastifyInstance) => {
     fastify.post('/nft/', mintNFT);
     fastify.get('/nfts/', getAllNFTs);
     fastify.get('/nft/:id', getNFT);
-    //fastify.post('/nft/transfer', transferNFT);
 };
 
 export default nftRoutes;
