+++ src/controllers/replyController.ts
@@ -1,22 +1,22 @@
 import axios from 'axios';
-import { connectToMongoDB } from '../db/dbConnections';
+import { connectToMongoDB } from './dbConnections';
 import mongoose from 'mongoose';
 import { userConversationSchema } from '../models/userConversation';
 import { SCROLL_CONFIG } from '../constants/networks';
 
-const mongoUrl = 'mongodb+srv://chatbot:p4tech@p4techsolutions.hvxfjoc.mongodb.net/chatterpay';
+const mongoUrl = process.env?.MONGO_URI_CHATTERPAY ?? "";
 
 export const sendTransferNotification = async (channel_user_id: string, from: string | null, amount: string, token: string) => {
     try {
         console.log("Sending notifications");
         const connection = await connectToMongoDB(mongoUrl);
-        
+
         // Creamos los modelos usando esta conexión específica
         const UserConversation = connection.model('user_conversations', userConversationSchema);
-        
+
         await updateUserConversationStatus(UserConversation, channel_user_id, "operator");
         console.log("Updated conversation to operator");
-        
+
         const payload: OperatorReplyPayload = {
             data_token: 'ddbe7f0e3d93447486efa9ef77954ae7',
             channel_user_id: channel_user_id,
@@ -25,7 +25,7 @@ export const sendTransferNotification = async (channel_user_id: string, from: st
         await sendOperatorReply(payload);
         console.log("Sent operator reply");
 
-		await updateUserConversationStatus(UserConversation, channel_user_id, "assistant");
+        await updateUserConversationStatus(UserConversation, channel_user_id, "assistant");
         console.log("Updated conversation to assistant");
     } catch (error) {
         console.error("Error in sendTransferNotification:", error);
@@ -34,14 +34,14 @@ export const sendTransferNotification = async (channel_user_id: string, from: st
 }
 
 export const sendSwapNotification = async (channel_user_id: string, token: string, amount: string, result: string, outputToken: string, transactionHash: string) => {
-    
+
     try {
         console.log("Sending notifications");
         const connection = await connectToMongoDB(mongoUrl);
-        
+
         // Creamos los modelos usando esta conexión específica
         const UserConversation = connection.model('user_conversations', userConversationSchema);
-        
+
         const payload: OperatorReplyPayload = {
             data_token: 'ddbe7f0e3d93447486efa9ef77954ae7',
             channel_user_id: channel_user_id,
@@ -62,10 +62,10 @@ export const sendMintNotification = async (channel_user_id: string, id: number) 
     try {
         console.log("Sending notifications");
         const connection = await connectToMongoDB(mongoUrl);
-        
+
         // Creamos los modelos usando esta conexión específica
         const UserConversation = connection.model('user_conversations', userConversationSchema);
-        
+
         const payload: OperatorReplyPayload = {
             data_token: 'ddbe7f0e3d93447486efa9ef77954ae7',
             channel_user_id: channel_user_id,
@@ -80,18 +80,18 @@ export const sendMintNotification = async (channel_user_id: string, id: number) 
     }
 }
 
-export const sendTransferNotification2 = async (channel_user_id: string, to:string | null, amount: string, token: string, txHash:string) => {
-    
+export const sendTransferNotification2 = async (channel_user_id: string, to: string | null, amount: string, token: string, txHash: string) => {
+
     try {
         console.log("Sending notifications");
         const connection = await connectToMongoDB(mongoUrl);
-        
+
         // Creamos los modelos usando esta conexión específica
         const UserConversation = connection.model('user_conversations', userConversationSchema);
-        
+
         await updateUserConversationStatus(UserConversation, channel_user_id, "operator");
         console.log("Updated conversation to operator");
-        
+
         const payload: OperatorReplyPayload = {
             data_token: 'ddbe7f0e3d93447486efa9ef77954ae7',
             channel_user_id: channel_user_id,
@@ -100,7 +100,7 @@ export const sendTransferNotification2 = async (channel_user_id: string, to:stri
         await sendOperatorReply(payload);
         console.log("Sent operator reply");
 
-		await updateUserConversationStatus(UserConversation, channel_user_id, "assistant");
+        await updateUserConversationStatus(UserConversation, channel_user_id, "assistant");
         console.log("Updated conversation to assistant");
     } catch (error) {
         console.error("Error in sendTransferNotification:", error);
@@ -122,23 +122,23 @@ const updateUserConversationStatus = async (UserConversation: mongoose.Model<any
 };
 
 interface OperatorReplyPayload {
-	data_token: string;
-	channel_user_id: string;
-	message: string;
+    data_token: string;
+    channel_user_id: string;
+    message: string;
 }
 
 const sendOperatorReply = async (payload: OperatorReplyPayload) => {
-	try {
-		const response = await axios.post('https://chatterpay-i7bji6tiqa-uc.a.run.app/chatbot/conversations/send-message', payload, {
-			headers: {
-				'Content-Type': 'application/json',
-				'Authorization': `Bearer ${payload.data_token}`
-			}
-		});
-		console.log('Respuesta de la API:', response.data);
-		return response.data;
-	} catch (error) {
-		console.error('Error al enviar la respuesta del operador:', error);
-		throw error;
-	}
+    try {
+        const response = await axios.post('https://chatterpay-i7bji6tiqa-uc.a.run.app/chatbot/conversations/send-message', payload, {
+            headers: {
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${payload.data_token}`
+            }
+        });
+        console.log('Respuesta de la API:', response.data);
+        return response.data;
+    } catch (error) {
+        console.error('Error al enviar la respuesta del operador:', error);
+        throw error;
+    }
 };
